Refactoring: * Rename files * Rename functions * Change structure of project * Add comments * Reduce max symbols of post to 7000
